requirements.txt
fastapi
uvicorn
torch 
transformers
pillow
pydicom 
opencv-python